==> 2_weather_data.ipynb <==
import warnings
warnings.filterwarnings("ignore")

import requests
import pandas as pd
import json
import os
from dotenv import load_dotenv

load_dotenv()

API_key = os.getenv("weather_api")

city = "London"
url = f"http://api.openweathermap.org/data/2.5/forecast?q={city}&appid={API_key}&units=metric"

# an example API call to London
weather_ex = requests.get(url)
weather_ex.status_code

weather_ex.json()["list"]

weatherex_df = pd.json_normalize(weather_ex.json()["list"])
weatherex_df.head()

weatherex_df = weatherex_df[["dt", "pop", "dt_txt", "main.temp", "main.humidity", "clouds.all", "wind.speed", "wind.gust"]]
weatherex_df.head()

# read the city dataframe to get the names of all cities we want weather data for
city_data = pd.read_csv("city_data.csv")
city_data.drop(columns = ["Unnamed: 0"], inplace = True)
city_data.head()

# convert the city column to a list so it can be used as an input of the function call
cities = city_data["city"].to_list()

# function that makes a call to the weather API and concatenate weather data from all cities of interest into
# one dataframe
def get_weather_data(cities):
    weather_list = []
    url = f"http://api.openweathermap.org/data/2.5/forecast?q=Boston&appid={API_key}&units=metric"
    test = requests.get(url)
    if test.status_code >= 200 and test.status_code <= 299:
        for city in cities:
            url = f"http://api.openweathermap.org/data/2.5/forecast?q={city}&appid={API_key}&units=metric"
            weather = requests.get(url)
            weather_df = pd.json_normalize(weather.json()["list"])
            weather_df["city"] = city
            weather_list.append(weather_df)
    else:
        return -1
    weather_combined = pd.concat(weather_list, ignore_index = True)
    return weather_combined

weather_data = get_weather_data(cities)
weather_data

# set dataframe to keep only the columns of interest
weather_data = weather_data[["pop", "dt_txt", "main.temp", "main.feels_like", "main.humidity", "clouds.all", 
                             "wind.speed", "wind.gust", "city"]]

# rename columns with more intuitive names and get rid of dots to avoid problems in MySQL
weather_data.rename(columns = {"pop": "precip_prob", 
                               "dt_txt": "forecast_time", 
                               "main.temp": "temperature",
                               "main.feels_like": "feels_like",
                               "main.humidity": "humidity", 
                               "clouds.all": "cloudiness", 
                               "wind.speed": "wind_speed", 
                               "wind.gust": "wind_gust",}, 
                    inplace = True)

# add the city_id column to the weather dataframe which will be used as the foreign key in MySQL
weather_data = weather_data.merge(city_data[["city_id", "city"]], how = "left", on = "city")

weather_data.sample(10)

weather_data.dtypes

weather_data["forecast_time"] = pd.to_datetime(weather_data["forecast_time"])

weather_data.to_csv("weather_data.csv")
