==> 4_gans_SQL_setup.ipynb <==
import pandas as pd
import pymysql
import sqlalchemy
import os
from dotenv import load_dotenv

city_df = pd.read_csv("city_data.csv")
weather_df = pd.read_csv("weather_data.csv")
airport_df = pd.read_csv("airport_data.csv")
flight_df = pd.read_csv("flight_data.csv")

city_df.drop(columns = ["Unnamed: 0"], inplace = True)
weather_df.drop(columns = ["Unnamed: 0"], inplace = True)
airport_df.drop(columns = ["Unnamed: 0"], inplace = True)
flight_df.drop(columns = ["Unnamed: 0"], inplace = True)

city_df.head()

weather_df.head()

airport_df.head()

flight_df.head()

load_dotenv()

schema = "gans"
host = "127.0.0.1"
user = "root"
password = os.getenv("MYSQL_PASSWORD")
port = 3306
con = f"mysql+pymysql://{user}:{password}@{host}:{port}/{schema}"

city_df.to_sql("cities", if_exists = "append", con = con, index = False)

weather_df.to_sql("weather", if_exists = "append", con = con, index = False)

airport_df.to_sql("airports", if_exists = "append", con = con, index = False)

flight_df.to_sql("flights", if_exists = "append", con = con, index = False)
